- baml_cffi.h/cbindgen.toml/bridge_cffi tests: canary's generated-header
  machinery is authoritative (supersedes our header_is_current.rs);
  header regenerated to include BAML_BRIDGE_LANGUAGE_CPP = 6
- bridge_cpp headers ported to the renamed ABI (Buffer -> BamlBuffer,
  uintptr_t -> size_t, language id via the enum constant)
- sdkgen_cpp ported to QualifiedTypeName accessors and the CodegenTy
  struct-variant shapes; Ty::EnumVariant widens to its enum
- alias identity is now preserved at use sites (#4048), so generated
  signatures spell the using-alias name
- cbindgen pinned =0.29.0 (canary), typescript sdk crate renames taken

+++ baml_language/sdks/cpp/bridge_cpp/include/baml/detail/registry.h
@@ -15,7 +15,7 @@
 
 extern "C" inline void baml_cpp_result_trampoline(uint32_t call_id,
                                                   const int8_t* content,
-                                                  uintptr_t length);
+                                                  size_t length);
 
 namespace baml {
 namespace detail {
@@ -84,11 +84,10 @@ class CallRegistry {
 
 extern "C" inline void baml_cpp_result_trampoline(uint32_t call_id,
                                                   const int8_t* content,
-                                                  uintptr_t length) {
+                                                  size_t length) {
   // No C++ exception may cross the C ABI (bridge contract).
   try {
-    baml::detail::CallRegistry::Instance().Complete(
-        call_id, content, static_cast<size_t>(length));
+    baml::detail::CallRegistry::Instance().Complete(call_id, content, length);
   } catch (...) {
   }
 }
